+++ lhcb/lhcb_client/src/main/java/eu/rethink/lhcb/client/util/Tuple.java
@@ -26,4 +26,24 @@ public class Tuple<X, Y> {
         this.x = x;
         this.y = y;
     }
+
+    @Override
+    public String toString() {
+        return "(" + x + "," + y + ")";
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (o == this)
+            return true;
+        else if (o instanceof Tuple) {
+            try {
+                return x.equals(((Tuple) o).x) && y.equals(((Tuple) o).y);
+            } catch (Exception e) {
+                return false;
+            }
+        } else {
+            return false;
+        }
+    }
 }